Run the insertion effects in double, locals included
Same narrowing as the send effects, in the seven insertion effects.  Upstream
holds their scalar state in JavaScript numbers and only the delay lines in
Float32Array; the port had the scalars in float.

Widening the struct fields alone did nothing, which is the useful part of this:
the process loops copy that state into float locals, so a double field was
narrowed straight back on the way in.  The phaser's feedback path shows it
best -- prevL is written back as (float)apL every sample and read into the next
sample's all-pass input, so the rounding compounds through an eight-stage
chain.  The locals, the LFO phase and rate, the shelf and mix and level
arithmetic, and the single-precision math calls are all widened now; the
all-pass histories stay float32, as upstream's Float32Arrays are.

phaser 10.68 -> 48.46 dB, and the 367-sample offset the comparator had been
reporting is gone -- that was it failing to align a signal this far out, not a
real delay.  auto_wah 18.15 -> 54.55, tremolo_efx_test 39.78 -> 74.01.  Nothing
else in the corpus moves, the grid does not move, and the four C tests pass.
Major verdicts fall to fourteen and moderate to one.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/insertion.h
@@ -29,9 +29,9 @@ typedef struct {
 typedef struct SS_InsertionProcessor {
 	uint32_t type; /* MSB<<8|LSB, e.g. 0x0100 = StereoEQ */
 
-	float send_level_to_reverb;
-	float send_level_to_chorus;
-	float send_level_to_delay;
+	double send_level_to_reverb;
+	double send_level_to_chorus;
+	double send_level_to_delay;
 
 	/**
 	 * Process one block. Adds to output buffers.
